pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
packages = [{ include = "github_commit_pipeline", from = "src" }]

[tool.poetry.dependencies]
python        = "^3.12"
PyGithub      = "^2.3"   # GitHub API
minio         = "^7.2"   # S3-like storage
sqlalchemy    = "^2.0"
duckdb-engine = "^0.10"
python-dotenv = "^1.0"
pyyaml = "^6.0"